pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]


[project]
name = "oor_benchmark"
version = "0.0.1"
description = "Code for benchmarking out-of-reference detection"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
authors = [
    {name = "Emma Dann"},
]
maintainers = [
    {name = "Emma Dann", email = "ed6@sanger.ac.uk"},
]
urls.Documentation = "https://oor_benchmark.readthedocs.io/"
urls.Source = "https://github.com/emdann/oor_benchmark"
urls.Home-page = "https://github.com/emdann/oor_benchmark"
dependencies = [
    "anndata",
    "scanpy",
    "scvi-tools"
    ]

[project.optional-dependencies]
dev = [
    # CLI for bumping the version number
    "bump2version",
    "pre-commit"
]
doc = [
    "sphinx>=4",
    "furo",
    "myst-parser",
    "sphinxcontrib-bibtex>=1.0.0",
    "scanpydoc[typehints]>=0.7.4",
    # For notebooks
    "nbsphinx",
    "ipykernel"
]
test = [
    "pytest",
    "pytest-cov",
]

[tool.coverage.run]
source = ["oor_benchmark"]
omit = [
    "**/test_*.py",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
xfail_strict = true
addopts = [
    "--import-mode=importlib",  # allow using test files with same name
]

[tool.isort]
include_trailing_comma = true
multi_line_output = 3
profile = "black"
skip_glob = ["docs/*"]

[tool.black]
line-length = 120
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.jupytext]
formats = "ipynb,md"
